2.3.1 test7 preFinish

+++ src/main/java/web/Dao/UserDao.java
@@ -8,11 +8,11 @@ public interface UserDao {
 
     List<User> findAll();
 
-    User findOne(int id);
+    User findOne(Long id);
 
     void save(User user);
 
-    void update(int id, User updatedUser);
+    void update(Long id, User updatedUser);
 
-    void delete(int id);
+    void delete(Long id);
 }
